Fix color roles calculation
